Load testing

+++ cypress/integration/tests/applicationinventory/application/load_testing.test.ts
@@ -1,11 +1,13 @@
 /// <reference types="cypress" />
 
-import { login } from "../../../../utils/utils";
+import { login, clickByText } from "../../../../utils/utils";
 import { Tag } from "../../../models/tags";
 import { ApplicationInventory } from "../../../models/applicationinventory/applicationinventory";
 import { Stakeholders } from "../../../models/stakeholders";
 import * as data from "../../../../utils/data_utils";
 import { BusinessServices } from "../../../models/businessservices";
+import { applicationinventory} from "../../../types/constants";
+import { navMenu } from "../../../views/menu.view";
 
 var stakeholdersList: Array<Stakeholders> = [];
 var stakeholdersNameList: Array<string> = [];
@@ -18,40 +20,31 @@ describe("Load testing", () => {
         // Perform login
         login();
 
-        // Create multiple bussiness services and tags
-        for (let i = 0; i < 100; i++) {
-            const stakeholder = new Stakeholders(data.getEmail(), data.getFullName());
-            stakeholder.create();
-            stakeholdersList.push(stakeholder);
-            stakeholdersNameList.push(stakeholder.name);
-        }
+        const stakeholder = new Stakeholders(data.getEmail(), data.getFullName());
+        stakeholder.create();
+        stakeholdersList.push(stakeholder);
+        stakeholdersNameList.push(stakeholder.name);
 
-        // Create new business service
-        for (let j = 0; j < 100; j++) {
-            const businessservice = new BusinessServices(
-                data.getCompanyName(),
-                data.getDescription(),
-                stakeholdersList[j].name
-            );
-            businessservice.create();
-            businessservicesList.push(businessservice);
-        }
-
-        // Create new tag
-        for (let k = 0; k < 100; k++) {
-            const tag = new Tag(data.getRandomWord(6), data.getExistingTagtype());
+        const businessservice = new BusinessServices(
+            data.getCompanyName(),
+            data.getDescription(),
+            stakeholdersList[0].name
+        );
+        businessservice.create();
+        businessservicesList.push(businessservice);
+    
+        const tag = new Tag(data.getRandomWord(6), data.getExistingTagtype());
             tag.create();
             tagList.push(tag);
-        }
 
         // Navigate to application inventory tab and create new application
-        for (let l = 0; l < 100; l++) {
+        for (let l = 0; l < 2; l++) {
             const application = new ApplicationInventory(
                 data.getAppName(),
                 data.getDescription(),
                 data.getDescription(),
-                businessservicesList[l].name,
-                [tagList[l].name]
+                businessservicesList[0].name,
+                [tagList[0].name]
             );
             application.create();
             applicationList.push(application);
@@ -67,38 +60,50 @@ describe("Load testing", () => {
         cy.intercept("POST", "/api/controls/business-service*").as("postBusinessService");
         cy.intercept("GET", "/api/controls/business-service*").as("getBusinessService");
 
-        // Interceptors
         cy.intercept("POST", "/api/controls/tag*").as("postTag");
         cy.intercept("GET", "/api/controls/tag*").as("getTag");
 
-        // Interceptors
         cy.intercept("GET", "/api/application-inventory/application*").as("getApplication");
     });
 
     after("Perform test data clean up", function () {
-        // Delete the applications and stakeholders created before the tests
+        // Delete stakeholders created before the tests
         stakeholdersList.forEach(function (stakeholder) {
             stakeholder.delete();
         });
-
-        // Delete application
-        applicationList.forEach(function (application) {
-            application.delete();
-        });
+       
         // Clean up business service and tags
-        businessservicesList[1].delete();
-        tagList[1].delete();
+        businessservicesList.forEach(function (stakeholder) {
+            stakeholder.delete();
+        });
+
+        tagList.forEach(function (stakeholder) {
+            stakeholder.delete();
+        });
+
+        // Delete all applications
+        clickByText(navMenu, applicationinventory);
+        cy.wait(2000);
+        applicationList.forEach(function (application) {
+            cy.get(".pf-c-table > tbody > tr")
+                .not(".pf-c-table__expandable-row")
+                .find("td[data-label=Name]")
+                .each(($rows) => {
+                    if ($rows.text() === application.name) application.delete();
+                });
+        });
     });
 
     it("Assess and Review all applications", function () {
-        for (let i = 0; i < 100; i++) {
+        for (let i = 0; i < 2; i++) {
             // Perform assessment of application
-            applicationList[i].perform_assessment("high", stakeholdersNameList);
+            var risk = data.getRandomRisk()
+            applicationList[i].perform_assessment(risk, stakeholdersNameList);
             cy.wait(2000);
             applicationList[i].is_assessed();
 
             // Perform application review
-            applicationList[i].perform_review("high");
+            applicationList[i].perform_review(risk);
             cy.wait(2000);
             applicationList[i].is_reviewed();
         }
